requirements.txt
fastapi
python-dotenv
uvicorn
aiohttp
plugp100
fastapi-cors
